Rename `XFeignFederationClient` to `XFederationFeignClient`...
