+++ common/federation/src/test/java/app/coronawarn/server/common/federation/client/DevelopmentFederationFeignHttpClientProviderSmokeTest.java
@@ -3,26 +3,22 @@ package app.coronawarn.server.common.federation.client;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import app.coronawarn.server.common.federation.client.config.FederationGatewayConfig;
-import app.coronawarn.server.common.federation.client.config.FederationGatewayConfig.Ssl;
-import java.io.File;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(properties = "spring.main.lazy-initialization=true", webEnvironment = WebEnvironment.RANDOM_PORT)
 @DirtiesContext
-class CloudFeignFederationHttpClientProviderSmokeTest {
+class DevelopmentFederationFeignHttpClientProviderSmokeTest {
+
+  @Autowired
+  private FederationGatewayConfig config;
 
   @Test
-  void testCanLoadKeystore() {
-    Ssl ssl = new Ssl();
-    ssl.setKeyStore(new File("../../docker-compose-test-secrets/ssl.p12"));
-    ssl.setKeyStorePass("");
-    FederationGatewayConfig config = new FederationGatewayConfig();
-    config.setConnectionPoolSize(1);
-    config.setSsl(ssl);
-    CloudFeignFederationHttpClientProvider cut = new CloudFeignFederationHttpClientProvider(config);
+  void testCanCreateHttpClient() {
+    DevelopmentFederationFeignHttpClientProvider cut = new DevelopmentFederationFeignHttpClientProvider(config);
     assertThat(cut.createFeignClient()).isNotNull();
   }
